docs(examples): make sure examples are functional again
google blocks automated navigation, now using duckduckgo
